Use system flux instead of an embedded flux
This just calls "flux" instead of a specific path - it has the neat
side-effect that instead of using WEAVE_GITOPS_FLUX_BIN_PATH, you can
just set your own flux bin path using the PATH variable.

Right now, there are only two things we do with this flux: creating a
secret, and wrapping `flux check` in `gitops check`.

The gitops check should end up removing the flux check.

The secret is part of accessing private git repos, which is called by
auth which is called by both upgrade and profiles.

There's still a flux installed by `make dependencies` - I've adapted
the auth tests to use the one installed by that, but as it's not added
to your path, it's not going to be used by the local CLI. Flux is now
a dependency I expect you to install yourself (will not document right
now, but note to self to add to the documentation #1532 is adding).