Remove un-needed ui makefile targets.
The `cmd/gitops/ui/run/dist/{,index.html}` targets create an empty
directory and file respectively and are only used in github workflows to
avoid git errors due to missing embed targets. This removes them in
favour of making it clear in the github actions that nothing is actually
being made.

This also re-orders the dependencies for the ui-lib target within the
makefile to make it clearer what lives with what